tried instlng datetime picker.. but not working

+++ client/app/stage/stage.controller.js
@@ -12,6 +12,12 @@ angular.module('myAppApp')
     	$scope.stage = response.stage;
         console.log($scope.stage);
         $scope.isFollowing = response.isFollowing;
+        if($scope.isFollowing){
+            $scope.follow='Following';
+        }
+        else{
+            $scope.follow='Follow';
+        }
     });
 
     $http.get('/api/posts/stage/'+id).success(function (response){
@@ -19,21 +25,31 @@ angular.module('myAppApp')
         $scope.posts = response;
         console.log($scope.posts);
     });
+
     $scope.follow = function(){
-        $http.get('/api/stages/'+id+'/addSubscriber').success(function (response){
-            console.log(response);
-            if(response.added==true){
-                $scope.isFollowing = true;
-            }
-        });
+        console.log('gdhs');
+        if(!$scope.isFollowing){
+            $http.get('/api/stages/'+id+'/addSubscriber').success(function (response){
+                console.log(response);
+                if(response.added==true){
+                    $scope.isFollowing = true;
+                    $scope.follow='Following';
+                }
+            });
+        }
+        else {
+            $http.delete('/api/stages/'+id+'/deleteSubscriber').success(function (response){
+                console.log(response);
+                if(response.removed==true){
+                    $scope.isFollowing = false;
+                    $scope.follow='Follow';
+                }
+            });
+        }
+        
+
+        
     }
-    $scope.unfollow = function(){
-        $http.delete('/api/stages/'+id+'/deleteSubscriber').success(function (response){
-            console.log(response);
-            if(response.removed==true){
-                $scope.isFollowing = false;
-            }
-        });
-    }
+   
 
   });
